+++ app/components/EnhancedTableToolbar.js
@@ -2,16 +2,11 @@ import React, { Component } from 'react';
 import classNames from 'classnames';
 import { withStyles } from '@material-ui/core/styles';
 import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-import IconButton from '@material-ui/core/IconButton';
 import Tooltip from '@material-ui/core/Tooltip';
-import InfoIcon from '@material-ui/icons/Info';
-import FilterListIcon from '@material-ui/icons/FilterList';
 import Fab from '@material-ui/core/Fab';
 import AddIcon from '@material-ui/icons/Add';
 import Menu from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
-import IntegrationAutosuggest from '../components/IntegrationAutosuggest';
 import { ux } from '../utils/ux';
 
 const toolbarStyles = theme => ({
@@ -42,9 +37,7 @@ type Props = {
   numSelected: number,
   enableEditContainer?: boolean,
   handleAddByFile?: () => {},
-  handleAddByFolder?: () => {},
-  getSuggestions?: () => {},
-  onAutosuggestInputChanged?: () => {}
+  handleAddByFolder?: () => {}
 };
 
 class EnhancedTableToolbar extends Component<Props> {
@@ -78,27 +71,9 @@ class EnhancedTableToolbar extends Component<Props> {
     const { anchorEl } = this.state;
     return (
       <Toolbar
-        className={classNames(classes.root, {
-          [classes.highlight]: numSelected > 0,
-        })}
+        className={classNames(classes.root)}
       >
-        <div className={classes.title}>
-          {numSelected > 0 ? (
-            <Typography color="inherit" variant="subtitle1">
-              {numSelected} selected
-            </Typography>
-          ) : (
-            <Typography variant="h6" id="tableTitle" />
-          )}
-        </div>
         <div className={classes.spacer} />
-        <div style={{ width: 700 }}>
-          {enableEditContainer ? (
-            <IntegrationAutosuggest
-              getSuggestions={this.props.getSuggestions}
-              onInputChanged={this.props.onAutosuggestInputChanged}
-            />) : null}
-        </div>
         <div className={classes.actions}>
           {handleAddByFile ? (
             <div>
@@ -127,21 +102,6 @@ class EnhancedTableToolbar extends Component<Props> {
           ) : null
         }
         </div>
-        <div className={classes.actions}>
-          {numSelected > 0 ? (
-            <Tooltip title="Summary">
-              <IconButton aria-label="Summary">
-                <InfoIcon />
-              </IconButton>
-            </Tooltip>
-          ) : (
-            <Tooltip title="Filter list">
-              <IconButton aria-label="Filter list">
-                <FilterListIcon />
-              </IconButton>
-            </Tooltip>
-          )}
-        </div>
       </Toolbar>
     );
   }
@@ -150,9 +110,7 @@ class EnhancedTableToolbar extends Component<Props> {
 EnhancedTableToolbar.defaultProps = {
   enableEditContainer: false,
   handleAddByFile: undefined,
-  handleAddByFolder: undefined,
-  getSuggestions: undefined,
-  onAutosuggestInputChanged: undefined
+  handleAddByFolder: undefined
 };
 
 export default withStyles(toolbarStyles)(EnhancedTableToolbar);